Extract a utility function to make it reusable
