+++ web_external/views/adapters/index.js
@@ -1,3 +1,5 @@
+import geo from 'geojs';
+
 import registry from './registry';
 import MapRepresentation from './MapRepresentation';
 import GeometryRepresentation from './GeometryRepresentation';
@@ -6,7 +8,6 @@ import WmsRepresentation from './WmsRepresentation';
 import ContourRepresentation from './ContourRepresentation';
 import ChoroplethRepresentation from './ChoroplethRepresentation';
 
-import geo from 'geojs';
 window.geo = geo;
 
 export {
@@ -17,4 +18,4 @@ export {
     WmsRepresentation,
     ContourRepresentation,
     ChoroplethRepresentation
-}
+};
